Created build environment
I started the basis for the build environment.  All of my current required libraries seems to be building and linking correctly.  The goal isn't to be highly portable, but this is working on 10.6.8 and I would expect that it should work on Linux/BSD.  It depends on libSIMDx86 and OpenMP.  I may use the smartptr from Boost.